+++ R/unpack_tarball.R
@@ -36,4 +36,3 @@ unpack_tarball <- function(pkg_tar,
 }
 
 
-