+++ tests/testthat/test-modify_description_file.R
@@ -304,4 +304,3 @@ test_that("insensitive case", {
 
 
 
-